### 2025-03-03 #7864 @editorialbot
[PRE REVIEW]: SideScanSonarEditor: A Python package for annotation of side-scan sonar data
**Submitting author:** @MichalMotylinski (<a href="http://orcid.org/0000-0002-4764-7130">Michal Motylinski</a>)
**Repository:** https://github.com/MichalMotylinski/SideScanSonarEditor
**Branch with paper.md** (empty if default branch): master
**Version:** 0.1.0

Thanks for submitting your paper to JOSS @Michal Motylinski. **Currently, there isn't a JOSS editor assigned** to your paper.

@Michal Motylinski if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission.
