### 2025-03-15 #7864 @crvernon
:wave: @michal - I am going to run this through scope review with our larger editorial board to ensure that it meets our substantial scholarly effort requirement in regards to our review of GUI applications.  I'll be back in touch ASAP.  Thanks.
