### 2025-06-13 #7864 @malmans2
Hi @crvernon,

I'm going to pass on this one, as I just finished reviewing another JOSS submission.
