### 2025-06-21 #8445 @editorialbot
[REVIEW]: SideScanSonarEditor: A Python package for annotation of side-scan sonar data
**Submitting author:** @MichalMotylinski (<a href="http://orcid.org/0000-0002-4764-7130">Michal Motylinski</a>)
**Repository:** https://github.com/MichalMotylinski/SideScanSonarEditor
**Branch with paper.md** (empty if default branch): master
**Version:** 1.0.0
