### 2025-06-25 #8445 @crvernon
@Digdgeo thanks!  You can now run the `@editorialbot generate my checklist` command to generate your checklist
