### 2025-06-26 #8445 @Digdgeo
Hi @MichalMotylinski,

I'm continuing with the review and have a question about authorship. I notice that all commits in the repository are from you, but the paper lists multiple authors.

Could you please clarify the contributions of the other authors? For example, did they contribute through:
- Scientific supervision or guidance
- Conceptual design
- Testing and feedback
- Funding acquisition
- Other forms of collaboration

This helps ensure the authorship appropriately reflects all contributions to the work.

Thanks!
