### 2025-07-05 MichalMotylinski/SideScanSonarEditor#5 @rossbar
DOC: Include demo/reference dataset
Related to openjournals/joss-reviews#8445

I am able to successfully follow the [installation instructions](https://github.com/MichalMotylinski/SideScanSonarEditor?tab=readme-ov-file#install-and-requirements). However, once I have the app installed there are no follow-up instructions on how to run/load a dataset. This makes the project difficult to evaluate.

IMO, the best resolution of this issue would be to point users to a reference XTF dataset on the internet; assuming there is an open data repository containing such data somewhere. If not, including a reference dataset in the repo is also an option, but IMO keeping source code and data separate is vastly preferable.
