### 2025-07-05 MichalMotylinski/SideScanSonarEditor#6 @rossbar
PKG: Non-source usage pattern not clear
Related to openjournals/joss-reviews#8445

There is a [published package on PyPi](https://pypi.org/project/SideScanSonarEditor/) which is great, however the project does not seem to be formatted as a Python package so there is not an obvious way to use the app from the PyPi package.

In other words, the only pattern with which I was able to actually run the app required downloading the source repo and running `python app.py`. There's nothing wrong with this, but it's confusing to have a Python package hosted on PyPi with no clear means of accessing the functionality (e.g. `from SideScanSonarEditor import app`)
