### 2025-07-05 MichalMotylinski/SideScanSonarEditor#7 @rossbar
Future work: consider a plug-in on top of other viewing application
Related to openjournals/joss-reviews#8445

First off, I just want to state that this issue/comment is by no means a blocker for the JOSS submission!

While reading the source code and the future work section of the manuscript, I formulated the impression that a project like this could benefit from building on top of other well-supported data visualization frameworks. Given the technology used here (Qt), something like [napari](https://github.com/napari/napari/) could make a lot of sense. The advantages of formulating this functionality as a napari plugin are two-fold, IMO:

 1. You can make use of very well-supported, intuitive functionality that is native to napari. For example, rather than having to engineer/support bespoke labeling code, you can use napari's instead, which already supports both mask-style (i.e. label images) and shape-style (i.e. arbitrary polygons) interactive labeling.
 2. The project would likely become more discoverable via association with a widely adopted ecosystem.

Again, I want to stress that *I am not suggesting you do this for the JOSS submission*! This is only something I wanted to raise for your consideration in supporting/maintaining this project!
