### 2025-07-05 MichalMotylinski/SideScanSonarEditor#8 @rossbar
TST: Testing and software sustainability
Related to openjournals/joss-reviews#8445

I didn't see any explicit testing for the package in the source repository. I am fully aware of how difficult it is to test GUI's, and do not think a UI unit testing framework is necessary (nor even beneficial) for the project.

However, I *do* think that at the very least installation testing would be useful. For example, adding a testing action that attempts to install the app on supported platforms against a range of e.g. Python versions. Note for example that the [README states](https://github.com/MichalMotylinski/SideScanSonarEditor?tab=readme-ov-file#install-and-requirements) that the app is tested on Python 3.9 (the support window for which ends in Oct. 2025). For users looking for sustainable software, the lack of explicit support/testing for modern Python distributions is likely to be a blocker (it would be for me, for example).

The simplest way (IMO) to address this issue would be to add a CI workflow to attempt to install the package from source with Python versions spanning the support window (which, according to the `pyproject.toml` is Python 3.9 through 3.13). Note that such testing would not guarantee that the app *works* as expected (unit tests would be valuable as well!) but at the very least would test that the app can be successfully installed.
