### 2025-07-11 #8445 @MichalMotylinski
@rossbar
Majority of the the datasets are just available in the GeoTiff format however this one will allow you to test the software:
https://www.marine-geo.org/tools/search/Files.php?data_set_uid=6159

I have just tested it myself and the only missing information is the Easting coordinate which might have been stripped so it will appear as 0.
