### 2025-07-27 8304ac7
Modified shape display to load points in the central part of each pixel. Removed snaping to grid.
