### 2025-07-27 MichalMotylinski/SideScanSonarEditor#5 @MichalMotylinski
I have added a sample XTF file in the project files since I cannot share the dataset used for research. I think this should enable testing of the software. Please let me know if there are any further issues with the dataset.
