### 2025-07-27 ab23ed8
Create main.yml
Added test for installation with different python versions.
