+++ .github/workflows/main.yml
@@ -1,0 +1,38 @@
+name: Test Package Installation
+
+on:
+  push:
+    branches: [ main ]
+  pull_request:
+    branches: [ main ]
+  workflow_dispatch:
+
+jobs:
+  install-test:
+    name: Python ${{ matrix.python-version }} install test
+    runs-on: ubuntu-latest
+
+    strategy:
+      matrix:
+        python-version: ["3.9", "3.10", "3.11", "3.12", "3.13"]
+
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v4
+
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v5
+        with:
+          python-version: ${{ matrix.python-version }}
+
+      - name: Upgrade pip and build dependencies
+        run: |
+          python -m pip install --upgrade pip setuptools wheel
+
+      - name: Install the package from source
+        run: |
+          pip install .
+
+      - name: Verify import
+        run: |
+          python -c "import SideScanSonarEditor"
