### 2025-07-27 05dec71
Updated code to allow running GUI from python script.
