### 2025-07-27 cd41194
Update main.yml
Updated run command.
