### 2025-07-27 e820614
Update main.yml
Testing just 3.9 version.
