### 2025-07-27 62d6148
Update main.yml
Added libegl1 installation.
