+++ .github/workflows/main.yml
@@ -20,6 +20,9 @@ jobs:
       - name: Checkout repository
         uses: actions/checkout@v4
 
+      - name: Install system dependencies for PyQt6
+        run: sudo apt-get update && sudo apt-get install -y libegl1 libgl1
+
       - name: Set up Python ${{ matrix.python-version }}
         uses: actions/setup-python@v5
         with:
