### 2025-07-27 a639f76
Update main.yml
Added dependencies installation.
