+++ .github/workflows/main.yml
@@ -9,19 +9,16 @@ on:
 
 jobs:
   install-test:
-    name: Python ${{ matrix.python-version }} install test
     runs-on: ubuntu-latest
-
-    strategy:
-      matrix:
-        python-version: ["3.9"]
-
     steps:
-      - name: Checkout repository
-        uses: actions/checkout@v4
+      - uses: actions/checkout@v4
 
-      - name: Install system dependencies for PyQt6
-        run: sudo apt-get update && sudo apt-get install -y libegl1 libgl1
+      - name: Install Qt system dependencies
+        run: |
+          sudo apt-get update
+          sudo apt-get install -y \
+            libegl1 libgl1 libxcb-xinerama0 libxcb-icccm4 libxcb-image0 \
+            libxcb-keysyms1 libxcb-render-util0 libxcb-xfixes0 libx11-xcb1 libxcb1 libxcb-render0 libxrender1
 
       - name: Set up Python ${{ matrix.python-version }}
         uses: actions/setup-python@v5
@@ -29,13 +26,10 @@ jobs:
           python-version: ${{ matrix.python-version }}
 
       - name: Upgrade pip and build dependencies
-        run: |
-          python -m pip install --upgrade pip setuptools wheel
+        run: python -m pip install --upgrade pip setuptools wheel
 
-      - name: Install the package from source
-        run: |
-          pip install .
+      - name: Install package
+        run: pip install -e .
 
-      - name: Verify import
-        run: |
-          python -c "from SideScanSonarEditor.app import main; main()"
+      - name: Run main()
+        run: python -c "from SideScanSonarEditor.app import main; main()"
