### 2025-07-27 3c888ce
Update main.yml
Additional dependecies installation.
