+++ .github/workflows/main.yml
@@ -17,7 +17,7 @@ jobs:
         run: |
           sudo apt-get update
           sudo apt-get install -y \
-            libegl1 libgl1 libxcb-xinerama0 libxcb-icccm4 libxcb-image0 \
+            libegl1 libgl1 libxcb-cursor0 libxcb-xinerama0 libxcb-icccm4 libxcb-image0 \
             libxcb-keysyms1 libxcb-render-util0 libxcb-xfixes0 libx11-xcb1 libxcb1 libxcb-render0 libxrender1
 
       - name: Set up Python ${{ matrix.python-version }}
