### 2025-07-27 4acf809
Update main.yml
Added dependencies and changed running to offscreen.
