### 2025-07-27 302f8a7
Update main.yml
Added headless check
