### 2025-07-27 3e5ff72
Update main.yml
Added different versions of python to test.
