### 2025-07-27 MichalMotylinski/SideScanSonarEditor#6 @MichalMotylinski
I have now added a possibility to run the GUI as a package PyPi package.
