### 2025-07-27 MichalMotylinski/SideScanSonarEditor#7 @MichalMotylinski
I think this is a very good idea. I am not sure how many changes would have to be made to [napari](https://github.com/napari/napari/) or other similar software since for side-scan sonar processing we need to process both channels separately (even though displayed in a single window). I will research the topic further because a more user friendly GUI (and more commonly used) would certainly benefit any potential users.
Thank you for bringing up this idea and [napari](https://github.com/napari/napari/) example.
