### 2025-07-27 080eb73
Update paper.md
Fixed bullet points and sentence about key features.
