### 2025-08-03 8b3b9f7
Updated project license.
