+++ pyproject.toml
@@ -9,7 +9,7 @@ description = "A Python package for editing side-scan sonar data."
 authors = [
     {name = "Michal Motylinski", email = "motylm66@gmail.com"}
 ]
-license = {text = "MIT"}
+license = {text = "LICENSE"}
 readme = "README.md"
 requires-python = ">=3.9"
 dependencies = [
