### 2025-08-04 9e4627e
Added authors contribution section
