### 2025-08-04 4ec2396
Removed checks from legacy code.
