+++ SideScanSonarEditor/app.py
@@ -781,106 +781,66 @@ class MyWindow(QMainWindow):
         if self.merged_image is None:
             return
 
-        try:
-            with open(os.path.join(self.input_filepath, self.tiles_filename), "r") as f:
-                old_tiles = json.load(f)
-        except:
-            old_tiles = {}
-        
         with open(os.path.join(self.input_filepath, self.tiles_filename), "w") as f:
             data = {}
             data["full_height"] = self.load_params["full_image_height"]
             data["full_width"] = self.load_params["full_image_width"]
             tiles = {}
-            i = 0
-
             for tile_data in self.canvas._tiles:
-                if tile_data == None:
-                    if str(i) in old_tiles.keys():
-                        tiles[i] = old_tiles[str(i)]
-                        i += 1
-                elif tile_data == "del":
-                    tiles[i] = "del"
-                    i += 1
-                else:
-                    tiles[i] = {"rectangle": [math.floor(tile_data["tiles"].rect().x()) * self.load_params["decimation"], math.floor((self.port_image.size[1] - math.floor(tile_data["tiles"].rect().y())) / self.load_params["stretch"]), tile_data["tiles"].rect().width() * self.load_params["decimation"], math.floor(math.floor(tile_data["tiles"].rect().height()) / self.load_params["stretch"])]}
-                    i += 1
+                if tile_data == "del":
+                    continue
 
-            # Remove polygons from dict 
-            for key in list(tiles.keys()):
-                if tiles[key] == "del":
-                    del tiles[key]
-
+                rect = tile_data["tiles"].rect()
+                tile_entry = {
+                    "rectangle": [
+                        math.floor(rect.x()) * self.load_params["decimation"],
+                        math.floor((self.port_image.size[1] - math.floor(rect.y())) / self.load_params["stretch"]),
+                        rect.width() * self.load_params["decimation"],
+                        math.floor(math.floor(rect.height()) / self.load_params["stretch"]),
+                    ]
+                }
+                tiles[len(tiles)] = tile_entry
             data["shapes"] = tiles
             json.dump(data, f, indent=4)
 
-        try:
-            with open(os.path.join(self.input_filepath, self.labels_filename), "r") as f:
-                old_polygons = json.load(f)
-        except:
-            old_polygons = {}
         
         with open(os.path.join(self.input_filepath, self.labels_filename), "w") as f:
             data = {}
             data["full_height"] = self.load_params["full_image_height"]
             data["full_width"] = self.load_params["full_image_width"]
-            new_polygons = self.canvas._polygons
+
+            new_polygons_raw = self.canvas._polygons
             polygons = {}
-            i = 0
-            
-            old_classes = {}
 
-            for polygon_data in new_polygons:
-                # If polygon not on the slice load it from file
-                # If the polygon was deleted then add "del" string for later removal from dict
-                # Any new or updated polygons add/update to the dict
-                if polygon_data == None:
-                    if str(i) in old_polygons["shapes"].keys():
-                        polygons[i] = old_polygons["shapes"][str(i)]
+            # Step 1: Build cleaned polygon dict (skip "del")
+            for i, polygon_data in enumerate(p for p in new_polygons_raw if p != "del"):
+                corners = []
+                for corner in polygon_data["polygon"]._polygon_corners:
+                    x = math.floor(corner[0]) * self.load_params["decimation"]
+                    y = math.floor(self.port_image.size[1] / self.load_params["stretch"] -
+                                math.floor(corner[1] / self.load_params["stretch"]))
+                    corners.append([x, y])
 
[…]
