+++ requirements.txt
@@ -1,4 +1,5 @@
 certifi==2025.1.31
+mock==5.2.0
 numpy==2.0.2
 opencv-python==4.11.0.86
 pillow==11.1.0
@@ -6,5 +7,7 @@ pyproj==3.6.1
 PyQt6==6.8.1
 PyQt6-Qt6==6.8.2
 PyQt6_sip==13.10.0
+pytest==8.4.1
+pytest-qt==4.5.0
 pyxtf==1.4.2
 scipy==1.13.1
