### 2025-08-04 168dc9f
Automated UI tests
