### 2025-08-04 928a593
Update ui_tests.yml
