+++ .github/workflows/ui_tests.yml
@@ -15,6 +15,14 @@ jobs:
       with:
         python-version: '3.11'
 
+    - name: Install Qt system dependencies
+      run: |
+          sudo apt-get update
+          sudo apt-get install -y \
+            libegl1 libgl1 libxcb-cursor0 libxcb-xinerama0 libxcb-icccm4 libxcb-image0 \
+            libxcb-keysyms1 libxcb-render-util0 libxcb-xfixes0 libxcb-shape0 libxcb-shm0 \
+            libxcb-sync1 libx11-xcb1 libxcb1 libxcb-render0 libxrender1 libxkbcommon-x11-0
+
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
