### 2025-08-04 8ff15b7
Added testing with X virtual framebuffer
