### 2025-08-04 f601059
Removed regular test
