### 2025-08-10 MichalMotylinski/SideScanSonarEditor#5 @rossbar
Thanks @MichalMotylinski that's helpful - I'd also recommend adding a link to the [other repository you mentioned in the other thread](https://github.com/openjournals/joss-reviews/issues/8445#issuecomment-3061402791) to your docs. This will allow users to test your software with "real" data!
