### 2025-08-10 MichalMotylinski/SideScanSonarEditor#9 @rossbar
Restrictive pinning in `requirements.txt` impedes installation
Coming back to this I just tried `pip instal -r requirements.txt` which resulted in source tarballs for e.g. numpy being downloaded in Python3.13. This is because the pins in `requirements.txt` are specified very tightly (down to patch versions).

Note however that the `pyproject.toml` does not have these pins, and therefore does not have these problems. My recommendation would be to get rid of `requirements.txt` since it seems like the full dependency footprint is already covered in the project specs. If for some reaseon you want to keep `requirements.txt`, I strongly encourage removing or at the very lease loosening the current pinning scheme to ensure that the software remains installable for future Python versions.
