### 2025-08-10 #8445 @rossbar
> please provide a short update here in this thread to how things are going. Thank you.

There has been a lot of progress since the last round of review, thanks @MichalMotylinski !

I too was able to download the [reference datasets](https://github.com/openjournals/joss-reviews/issues/8445#issuecomment-3061402791) to test out the software. The UI features work as expected. I've checked off the relevant boxes from the checklist accordingly. A few other bullets I'd like to highlight:

>  **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?

This is very clearly covered in the paper, but is not as robustly expressed in the README which serves as the documentation for the project. An [additional sentence or two](https://github.com/MichalMotylinski/SideScanSonarEditor/blob/3da5c88d6f96566a7d2ae0a20486896657df0863/README.md?plain=1#L4) or perhaps a link to the paper would resolve this IMO. This is a minor suggestion, not a blocker, so I've checked the bullet regardless!

>  **Automated tests:** Are there automated tests or manual steps described so that the functionality of the software can be verified?

A test suite and CI/CD workflows have been added so I consider this addressed. The one thing I'd highlight is to add something to the documentation (perhaps the contributor guide or a developer section) about how to run the test suite.

>  **Quality of writing:** Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?

Formatting/wording issues were sufficiently addressed in https://github.com/MichalMotylinski/SideScanSonarEditor/pull/3

---

My review checklist is complete. I still have some minor usability/sustainability concerns highlighted in https://github.com/MichalMotylinski/SideScanSonarEditor/issues/7, https://github.com/MichalMotylinski/SideScanSonarEditor/issues/9, and a minor suggestion in https://github.com/MichalMotylinski/SideScanSonarEditor/issues/5 but would be happy to see them addressed asynchronously; i.e. I don't consider them blockers for this review.
