### 2025-08-17 6e3afab
Update README.md
Added launching instructions.
