### 2025-08-17 7b94e5e
Update README.md
Added statement of need to readme.
