+++ README.md
@@ -2,6 +2,7 @@
 
 # SideScanSonarEditor
 SideScanSonarEditor is an open-source application designed for annotation of side-scan sonar data.
+The interpretation of side-scan sonar data is typically conducted using expensive proprietary software and annotation for the purpoeses training ML solutions requires use of multiple software applications and conversion between various formats. No open-source solutions currently exist that streamline this process, leaving researchers with time-consuming and fragmented tools. SideScanSonarEditor addresses this need by providing a free, open-source platform that simplifies visualization, annotation, and dataset preparation, enabling efficient integration with computer vision models.
 
 ## Install and Requirements
 The SideScanSonarEditor tool has been built and tested using Python 3.9 and requires following dependencies:
