### 2025-08-17 928f149
Update ui_tests.yml
Updated test to use .toml file
