### 2025-08-17 6bc4d67
Update pyproject.toml
Added test dependencies.
