### 2025-08-17 928c88b
Update ui_tests.yml
Updated test pipeline to utilise the optional dependencies.
