### 2025-08-17 MichalMotylinski/SideScanSonarEditor#9 @MichalMotylinski
Removed requirements.txt and changed testing pipeline to use pyproject.toml.
