### 2025-08-17 MichalMotylinski/SideScanSonarEditor#6 @MichalMotylinski
Added appropriate section in the README.md with instructions on how to launch the app.
