### 2025-08-17 #8445 @MichalMotylinski
Thank you very much @rossbar  and @Digdgeo.
I have corrected mentioned issue with dependencies installation and modified testing pipeline accordingly.  I also added a subsection with instructions on how to run the tests in CONTRIBUTING.md.
A brief statement of need and instructions on how to run the app have been added to README.md
