### 2025-09-03 #8445 @crvernon
@MichalMotylinski your in-text citations and the bib file are not being referenced correctly in the paper.  Please see the JOSS submission [docs](https://joss.readthedocs.io/en/latest/example_paper.html) for instructions on how to do this.  Let me know if you have any questions and when you finish and I'll take a look.

FYI... you can compile the paper again by commenting the following:

`@editorialbot generate pdf`
