### 2025-09-15 7e69f00
Fixed bibliography and reference display.
