+++ manuscript/paper.md
@@ -32,7 +32,7 @@ bibliography: bibliography.bib
 Side-scan sonars are widely used instruments for high-resolution mapping of the seafloor, utilizing acoustic signals to generate detailed imagery. These systems play a crucial role in marine research, habitat assessment, underwater archaeology, and infrastructure inspections, providing valuable insights into seabed features and composition.
 
 Annotation of the targets is typically done using specialised software which does not produce results allowing for quick conversion to formats commonly used to train machine learning models.
-Our software `SideScanSonarEditor` is a free and open-source Python package which allows the user to read sonar data from XTF files, display it as images and allow further processing (XTF FILE FORMAT, n.d.).
+Our software `SideScanSonarEditor` is a free and open-source Python package which allows the user to read sonar data from XTF files, display it as images and allow further processing [@xtf].
 The key features include the ability to fully analyze sonar imagery, easily manipulate it, and annotate objects of interest.
 The tool additionally allows for drawing of polygon shapes as well as rectangular shapes which coordinates are used to crop smaller image tiles. Together they can be used in the process of dataset creation for further sonar analysis or computer vision tasks such as object detection or segmentation.
 (\autoref{fig:overview}).
@@ -49,19 +49,19 @@ The tool additionally allows for drawing of polygon shapes as well as rectangula
 
  - Map range – Minimum and maximum mapping ranges are by default automatically calculated based on the intensity input data. Both can be manually modified to change the representation of the features for easier interpretation.
 
- - Slant range correction - apply slant rage correction to compensate for the geometric distortion of the return signal. The corrected image represents true seafloor distances allowing for better alignment with the navigation data (Chang et al., 2010).
+ - Slant range correction - apply slant rage correction to compensate for the geometric distortion of the return signal. The corrected image represents true seafloor distances allowing for better alignment with the navigation data [@chang:2010].
 
 ![SideScanSonarEditor app \label{fig:overview}](overview.png)
 
 # Statement of Need
 
-Side-scan sonar is a commonly used instrument for mapping of the seabed. These devices operate by being towed behind a vessel, emitting sonar pulses that reflect off the seabed to create detailed images. The raw data collected by the sonar consists of overlapping representations of the seabed that must be processed before they can be used for machine learning applications, such as target recognition and labeling.
+Side-scan sonar is a commonly used instrument for mapping of the seabed. These devices operate by being towed behind a vessel, emitting sonar pulses that reflect off the seabed to create detailed images. The raw data collected by the sonar consists of overlapping representations of the seabed that must be processed before they can be used for machine learning applications, such as target recognition and labeling [@Motylinski:2025].
 
-Interpreting side-scan sonar data from sonar typically requires expensive proprietary software designed for in-depth analysis and post-processing. However, these software packages do not facilitate the extraction of imagery and annotations in a format suitable for training computer vision models or further processing. To train an automatic target recognition model, users must manually crop imagery in SonarWiz (or similar closed-source software) before annotating it using LabelMe or other standard tools (Wada, 2024). This process is highly time-consuming, requiring manual visual analysis of sonar swaths in SonarWiz or EdgeTech Discover to identify and crop relevant areas (Chesapeake Technology - Makers of SonarWiz, n.d.; EdgeTech, n.d.). After cropping, the images must then be manually re-analyzed and labeled in another software such as `LabelMe', adding further to the labor-intensive workflow.
[…]
