+++ manuscript/bibliography.bib
@@ -1,5 +1,5 @@
 
-@misc{sture_oysstupyxtf_2025,
+@misc{oysstupyxtf:2025,
 	title = {oysstu/pyxtf},
 	copyright = {MIT},
 	url = {https://github.com/oysstu/pyxtf},
@@ -11,7 +11,7 @@
 	note = {original-date: 2016-06-29T09:29:56Z},
 }
 
-@misc{noauthor_xtf_nodate,
+@misc{xtf,
 	title = {{XTF} file format information},
 	url = {https://www.exail.com/resources/knowledge-center/xtf-file-format-information},
 	abstract = {Everything you need to know about the XTF file format for data processing.},
@@ -20,7 +20,7 @@
 	journal = {Exail},
 }
 
-@article{chang_sidescan_2010,
+@article{chang:2010,
 	title = {{SIDESCAN} {SONAR} {IMAGE} {PROCESSING}:{CORRECTING} {BRIGHTNESS} {VARIATION} {AND} {PATCHING} {GAPS}},
 	volume = {18},
 	issn = {2709-6998},
@@ -34,13 +34,13 @@
 	year = {2010},
 }
 
-@misc{noauthor_chesapeake_nodate,
+@misc{chesapeake,
 	title = {Chesapeake {Technology} - {Makers} of {SonarWiz}},
 	url = {https://chesapeaketech.com/},
 	urldate = {2024-12-06},
 }
 
-@misc{wada_labelme_2024,
+@misc{wada:2024,
 	title = {Labelme: {Image} {Polygonal} {Annotation} with {Python}},
 	shorttitle = {Labelme},
 	url = {https://github.com/wkentaro/labelme},
@@ -52,7 +52,7 @@
 	doi = {10.5281/zenodo.5711226},
 }
 
-@article{lin_microsoft_2014,
+@article{lin:2014,
 	title = {Microsoft {COCO}: {Common} {Objects} in {Context}},
 	volume = {abs/1405.0312},
 	url = {http://arxiv.org/abs/1405.0312},
@@ -62,7 +62,22 @@
 	note = {arXiv: 1405.0312},
 }
 
-@misc{noauthor_edgetech_nodate,
+@article{Motylinski:2025,
+doi = {10.1088/1361-6501/ad99f1},
+url = {https://dx.doi.org/10.1088/1361-6501/ad99f1},
+year = {2024},
+month = {dec},
+publisher = {IOP Publishing},
+volume = {36},
+number = {1},
+pages = {015435},
+author = {Motylinski, Michal and Plater, Andrew J and Higham, Jonathan E},
+title = {Computer vision methods for side scan sonar imagery},
+journal = {Measurement Science and Technology},
+abstract = {Side-scan sonar (SSS) data presents a significant challenge to analysts due to the size of the material that requires processing. Automatic recognition systems can help in the interpretation and filtration of the data, reducing the time needed for analysis. Deep learning architectures have been employed for computer vision tasks in the underwater domain. The objective of this work is to review current deep learning methods for automatic image classification, object detection, semantic segmentation, and instance segmentation through statistical analysis of published studies. Further, this work includes an inventory of publicly available SSS data followed by a discussion on augmentation methods applied to SSS imagery. The study discusses current challenges in the automatic underwater target detection field, highlights main gaps within the domain and identifies potential research directions. Our proposal includes the creation of a standardised benchmark and the use of a common set of metrics allowing for consistent evaluation of future research works.}
+}
+
+@misc{edgetech,
 	title = {{EdgeTech}},
 	url = {https://www.edgetech.com/},
 	urldate = {2023-05-02},
