### 2025-09-15 f20f416
Fixed double citation.
