+++ manuscript/paper.md
@@ -57,7 +57,7 @@ The tool additionally allows for drawing of polygon shapes as well as rectangula
 
 Side-scan sonar is a commonly used instrument for mapping of the seabed. These devices operate by being towed behind a vessel, emitting sonar pulses that reflect off the seabed to create detailed images. The raw data collected by the sonar consists of overlapping representations of the seabed that must be processed before they can be used for machine learning applications, such as target recognition and labeling [@Motylinski:2025].
 
-Interpreting side-scan sonar data from sonar typically requires expensive proprietary software designed for in-depth analysis and post-processing. However, these software packages do not facilitate the extraction of imagery and annotations in a format suitable for training computer vision models or further processing [@lin:2014]. To train an automatic target recognition model, users must manually crop imagery in SonarWiz (or similar closed-source software) before annotating it using LabelMe or other standard tools [@wada:2024]. This process is highly time-consuming, requiring manual visual analysis of sonar swaths in SonarWiz or EdgeTech Discover to identify and crop relevant areas [@chesapeake, @edgetech]. After cropping, the images must then be manually re-analyzed and labeled in another software such as `LabelMe', adding further to the labor-intensive workflow.
+Interpreting side-scan sonar data from sonar typically requires expensive proprietary software designed for in-depth analysis and post-processing. However, these software packages do not facilitate the extraction of imagery and annotations in a format suitable for training computer vision models or further processing [@lin:2014]. To train an automatic target recognition model, users must manually crop imagery in SonarWiz (or similar closed-source software) before annotating it using LabelMe or other standard tools [@wada:2024]. This process is highly time-consuming, requiring manual visual analysis of sonar swaths in SonarWiz or EdgeTech Discover to identify and crop relevant areas [@chesapeake; @edgetech]. After cropping, the images must then be manually re-analyzed and labeled in another software such as `LabelMe', adding further to the labor-intensive workflow.
 
 To the best of our knowledge, there is currently no open-source software available for viewing, manipulating, and annotating side-scan sonar files. Even commercial software does not provide a streamlined workflow optimized for time efficiency. Our open-source and free-to-use software, SideScanSonarEditor, significantly simplifies and accelerates the annotation process for sonar data. The output format is designed for easy analysis, further processing, and seamless integration with object detection or segmentation models [@lin:2014].
 
