### 2025-10-01 92906d8
Update bibliography.bib
Added missing dates to citations.
