+++ manuscript/bibliography.bib
@@ -11,7 +11,7 @@
 	note = {original-date: 2016-06-29T09:29:56Z},
 }
 
-@misc{xtf,
+@misc{xtf:2025,
 	title = {{XTF} file format information},
 	url = {https://www.exail.com/resources/knowledge-center/xtf-file-format-information},
 	abstract = {Everything you need to know about the XTF file format for data processing.},
@@ -34,10 +34,10 @@
 	year = {2010},
 }
 
-@misc{chesapeake,
+@misc{chesapeake:2025,
 	title = {Chesapeake {Technology} - {Makers} of {SonarWiz}},
 	url = {https://chesapeaketech.com/},
-	urldate = {2024-12-06},
+	urldate = {2025-10-02},
 }
 
 @misc{wada:2024,
@@ -77,8 +77,8 @@ journal = {Measurement Science and Technology},
 abstract = {Side-scan sonar (SSS) data presents a significant challenge to analysts due to the size of the material that requires processing. Automatic recognition systems can help in the interpretation and filtration of the data, reducing the time needed for analysis. Deep learning architectures have been employed for computer vision tasks in the underwater domain. The objective of this work is to review current deep learning methods for automatic image classification, object detection, semantic segmentation, and instance segmentation through statistical analysis of published studies. Further, this work includes an inventory of publicly available SSS data followed by a discussion on augmentation methods applied to SSS imagery. The study discusses current challenges in the automatic underwater target detection field, highlights main gaps within the domain and identifies potential research directions. Our proposal includes the creation of a standardised benchmark and the use of a common set of metrics allowing for consistent evaluation of future research works.}
 }
 
-@misc{edgetech,
+@misc{edgetech:2025,
 	title = {{EdgeTech}},
 	url = {https://www.edgetech.com/},
-	urldate = {2023-05-02},
+	urldate = {2025-10-02},
 }
