### 2025-10-01 26c98ec
Update paper.md
Removed unnecessary pagebreak.
