### 2025-10-01 b714205
Update paper.md
Fixed citations
