### 2025-10-02 eaf3595
Update bibliography.bib
Corrected bibliography entries.
