+++ manuscript/bibliography.bib
@@ -6,50 +6,51 @@
 	abstract = {A python library for reading and writing eXtended Triton Format (XTF) files},
 	urldate = {2025-02-25},
 	author = {Sture, Øystein},
-	month = feb,
 	year = {2025},
-	note = {original-date: 2016-06-29T09:29:56Z},
+	note = {original-date: 2016-06-29}
 }
 
 @misc{xtf:2025,
 	title = {{XTF} file format information},
 	url = {https://www.exail.com/resources/knowledge-center/xtf-file-format-information},
-	abstract = {Everything you need to know about the XTF file format for data processing.},
+	abstract = {Everything you need to know about the XTF file format for data processing},
 	language = {en},
 	urldate = {2025-02-25},
-	journal = {Exail},
+	year = {2025},
+	journal = {Exail}
 }
 
 @article{chang:2010,
 	title = {{SIDESCAN} {SONAR} {IMAGE} {PROCESSING}:{CORRECTING} {BRIGHTNESS} {VARIATION} {AND} {PATCHING} {GAPS}},
 	volume = {18},
+	number = {6},
 	issn = {2709-6998},
 	shorttitle = {{SIDESCAN} {SONAR} {IMAGE} {PROCESSING}},
 	url = {https://jmstt.ntou.edu.tw/journal/vol18/iss6/1},
 	doi = {10.51400/2709-6998.1935},
-	number = {6},
 	journal = {Journal of Marine Science and Technology},
 	author = {Chang, Yet-Chung and Hsu, Shu-Kun and Tsai, Ching-Hui},
 	month = dec,
-	year = {2010},
+	year = {2010}
 }
 
 @misc{chesapeake:2025,
 	title = {Chesapeake {Technology} - {Makers} of {SonarWiz}},
 	url = {https://chesapeaketech.com/},
 	urldate = {2025-10-02},
+	year = {2025}
 }
 
 @misc{wada:2024,
 	title = {Labelme: {Image} {Polygonal} {Annotation} with {Python}},
 	shorttitle = {Labelme},
 	url = {https://github.com/wkentaro/labelme},
-	abstract = {Image Polygonal Annotation with Python (polygon, rectangle, circle, line, point and image-level flag annotation).},
+	abstract = {Image Polygonal Annotation with Python (polygon, rectangle, circle, line, point and image-level flag annotation)},
 	urldate = {2024-11-15},
 	author = {Wada, Kentaro},
 	month = nov,
 	year = {2024},
-	doi = {10.5281/zenodo.5711226},
+	doi = {10.5281/zenodo.5711226}
 }
 
 @article{lin:2014,
@@ -59,26 +60,27 @@
 	journal = {CoRR},
 	author = {Lin, Tsung-Yi and Maire, Michael and Belongie, Serge J and Bourdev, Lubomir D and Girshick, Ross B and Hays, James and Perona, Pietro and Ramanan, Deva and Doll'a r, Piotr and Zitnick, C Lawrence},
 	year = {2014},
-	note = {arXiv: 1405.0312},
+	note = {arXiv: 1405.0312}
 }
 
 @article{Motylinski:2025,
-doi = {10.1088/1361-6501/ad99f1},
-url = {https://dx.doi.org/10.1088/1361-6501/ad99f1},
-year = {2024},
-month = {dec},
-publisher = {IOP Publishing},
-volume = {36},
-number = {1},
-pages = {015435},
-author = {Motylinski, Michal and Plater, Andrew J and Higham, Jonathan E},
-title = {Computer vision methods for side scan sonar imagery},
-journal = {Measurement Science and Technology},
[…]
