### 2025-10-02 e79dd9e
Update paper.md
Small visual and typos corrections.
