+++ manuscript/paper.md
@@ -20,7 +20,7 @@ authors:
     equal-contrib: false
     affiliation: 1
 affiliations:
- - name: Department Geography and Planning, School of Environmental Sciences, University of Liverpool, Liverpool, UK
+ - name: Department of Geography and Planning, School of Environmental Sciences, University of Liverpool, Liverpool, UK
    index: 1
 
 date: 26 February 2025
@@ -49,7 +49,7 @@ The tool additionally allows for drawing of polygon shapes as well as rectangula
 
  - Map range – Minimum and maximum mapping ranges are by default automatically calculated based on the intensity input data. Both can be manually modified to change the representation of the features for easier interpretation.
 
- - Slant range correction - apply slant rage correction to compensate for the geometric distortion of the return signal. The corrected image represents true seafloor distances allowing for better alignment with the navigation data [@chang:2010].
+ - Slant range correction - apply slant range correction to compensate for the geometric distortion of the return signal. The corrected image represents true seafloor distances allowing for better alignment with the navigation data [@chang:2010].
 
 ![SideScanSonarEditor app \label{fig:overview}](overview.png)
 
@@ -57,17 +57,17 @@ The tool additionally allows for drawing of polygon shapes as well as rectangula
 
 Side-scan sonar is a commonly used instrument for mapping of the seabed. These devices operate by being towed behind a vessel, emitting sonar pulses that reflect off the seabed to create detailed images. The raw data collected by the sonar consists of overlapping representations of the seabed that must be processed before they can be used for machine learning applications, such as target recognition and labeling [@Motylinski:2025].
 
-Interpreting side-scan sonar data from sonar typically requires expensive proprietary software designed for in-depth analysis and post-processing. However, these software packages do not facilitate the extraction of imagery and annotations in a format suitable for training computer vision models or further processing [@lin:2014]. To train an automatic target recognition model, users must manually crop imagery in SonarWiz (or similar closed-source software) before annotating it using LabelMe or other standard tools [@wada:2024]. This process is highly time-consuming, requiring manual visual analysis of sonar swaths in SonarWiz or EdgeTech Discover to identify and crop relevant areas [@chesapeake:2025; @edgetech:2025]. After cropping, the images must then be manually re-analyzed and labeled in another software such as `LabelMe', adding further to the labor-intensive workflow.
+Interpreting side-scan sonar data from sonar typically requires expensive proprietary software designed for in-depth analysis and post-processing. However, these software packages do not facilitate the extraction of imagery and annotations in a format suitable for training computer vision models or further processing [@lin:2014]. To train an automatic target recognition model, users must manually crop imagery in `SonarWiz` (or similar closed-source software) before annotating it using `LabelMe` or other standard tools [@wada:2024]. This process is highly time-consuming, requiring manual visual analysis of sonar swaths in `SonarWiz` or `EdgeTech Discover` to identify and crop relevant areas [@chesapeake:2025; @edgetech:2025]. After cropping, the images must then be manually re-analyzed and labeled in another software such as `LabelMe`, adding further to the labor-intensive workflow.
 
-To the best of our knowledge, there is currently no open-source software available for viewing, manipulating, and annotating side-scan sonar files. Even commercial software does not provide a streamlined workflow optimized for time efficiency. Our open-source and free-to-use software, SideScanSonarEditor, significantly simplifies and accelerates the annotation process for sonar data. The output format is designed for easy analysis, further processing, and seamless integration with object detection or segmentation models [@lin:2014].
[…]
