### 2025-10-02 #8445 @MichalMotylinski
@crvernon Thank you very much. Please find corrected article proof below. I have applied proposed changes and corrected a few typos and formatting issues.
