### 2025-10-22 0a543c7
Update paper.md
Removed backticks from title
