### 2025-10-22 c760f21
Update pyproject.toml
New version
