### 2025-10-22 6defc39
Update pyproject.toml
First full release.
