### 2025-10-22 #8445 @MichalMotylinski
@crvernon
Version: 1.0.0
DOI: https://doi.org/10.5281/zenodo.17411077
